refactor(glm5): split call parsing responsibilities
Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
